% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdtemplate.R
\name{get_logo_url}
\alias{get_logo_url}
\title{Return RStudio Logo URL}
\usage{
get_logo_url(color = "default")
}
\arguments{
\item{color}{Specifies which logo, one of default, dark, white, or old.}
}
\value{
URL for the logo image
}
\description{
Return RStudio Logo URL
}
