% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes.R
\name{get_rstudio_cols}
\alias{get_rstudio_cols}
\title{Return RStudio Color Values}
\usage{
get_rstudio_cols(...)
}
\arguments{
\item{...}{Pass in the names of the desired colors. To see available color
names, run the function with no arguments.}
}
\description{
Return RStudio Color Values
}
