% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-connectors.R
\name{get_sales_data}
\alias{get_sales_data}
\title{Easily Connect to RStudio SFDC Data}
\usage{
get_sales_data(token = rstudioapi::askForSecret("SFDCtoken",
  "Enter SFDC token, it will be securely cached!", "SalesForce"),
  pwd = rstudioapi::askForSecret("SFDCpassword",
  "Enter SFDC Password, it will be securely cached!", "SalesForce"), uid)
}
\arguments{
\item{token}{By default, we'll use RStudio to prompt and store the SFDC token}

\item{pwd}{By default, we'll use RStudio to prompt and store the SFDC token}

\item{uid}{Your SFDC username}
}
\value{
An odbc connection object, see db.rstudio.com
}
\description{
Easily Connect to RStudio SFDC Data
}
