% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes.R
\name{scale_color_rstudio}
\alias{scale_color_rstudio}
\title{RStudio Color Scheme}
\usage{
scale_color_rstudio(palette = "main", discrete = TRUE, ...)
}
\arguments{
\item{palette}{One of main, grey, or cool}

\item{discrete}{Whether or not the color aesthetic is discrete or numeric.}

\item{...}{Other arguments to \code(discrete_scale) or
\code(scale_color_gradient)}
}
\description{
RStudio Color Scheme
}
