% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes.R
\name{scale_fill_rstudio}
\alias{scale_fill_rstudio}
\title{RStudio Fill Scheme}
\usage{
scale_fill_rstudio(palette = "main", discrete = TRUE, ...)
}
\arguments{
\item{palette}{One of main, grey, or cool}

\item{discrete}{Whether or not the fill aesthetic is discrete or numeric.}

\item{...}{Other arguments to \code(discrete_scale) or
\code(scale_fill_gradient)}
}
\description{
RStudio Fill Scheme
}
